requirements.txt
# Predictive Analytics System - Dependencies
# Core Data Science
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Model Persistence
joblib>=1.3.0

# Progress Bars
tqdm>=4.65.0
